pyproject.toml
[project]
name = "instagram-fact-checker"
version = "0.1.0"
description = "Scrape Instagram posts and fact-check them using AI vision, web search, and LLM reasoning"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.10.0",
    "fastapi>=0.104.0",
    "groq>=1.0.0",
    "instaloader>=4.15",
    "langchain>=0.3",
    "langchain-groq>=0.2",
    "langchain-community>=0.3",
    "langgraph>=0.2",
    "lxml>=6.0.2",
    "pillow>=12.1.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]
